requirements.txt
# requirements.txt — Sistema Gestión Clínica Salud Total
# Versiones fijadas para reproducibilidad en producción

# ── Framework principal ─────────────────────────────────────
streamlit==1.35.0

# ── Base de datos ───────────────────────────────────────────
SQLAlchemy==2.0.30
psycopg2-binary==2.9.9

# ── Autenticación ───────────────────────────────────────────
bcrypt==4.1.3
PyJWT==2.8.0
python-dotenv==1.0.1

# ── Gráficos ────────────────────────────────────────────────
plotly==5.22.0
altair==5.3.0

# ── Reportes PDF ────────────────────────────────────────────
reportlab==4.2.0

# ── Análisis de datos ───────────────────────────────────────
pandas==2.2.2
numpy==1.26.4

# ── Utilidades ──────────────────────────────────────────────
Pillow==10.3.0
python-dateutil==2.9.0